Extract resolvers
